+++ CHANGELOG.md
@@ -11,7 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update client for `cds-notification` service to support `arbitraryEmailAddress` parameter used in `>= v0.13.0`
+- Client `NotificationService` implementing the previous version of the interface (`NotificationV4`) is renamed to `NotificationServiceLegacyV4`.
+
 ### Deprecated
+
+- Deprecate client interface `NotificationV4`.
+    Migrate from `NotificationV4` to `NotificationV5` by setting both `Arbitrary...` fields to empty strings or switch to `NotificationServiceLegacyV4`.
 
 ### Removed